Add episode sampler for lyrics data
This commit adds an episode sampler for the lyrics data as a class
EpisodeSampler. The EpisodeSampler splits the dataset into three subsets
for training, validation, and testing. The information for the split can
be persisted and loaded for future runs with the same train/val/test
split.

Each episode randomly samples a batch of artists from the dataset. For
each artist, query and support batches of songs are sampled from the
dataset. If an artist does not have enough songs to fill the query and
support batches, they are skipped.

The lyrics are read from a directory structure following the pattern
`./[root dir]/[artist]/[song]`. After a song has been read, it is parsed
into a series of "words" or "tokens" (currently just by splitting on
whitespace). Each word is assigned an integer ID (which is optionally
persisted to a file). The sequence of word IDs is then transformed into
a numpy array, of a configurable length (zero-padded if the song is
shorter than the configured length).

As each song is numpy-ified, they are (optionally) cached in memory.
From benchmarks on my laptop, I've seen that this caching offers an
order of magnitude performance improvement for each call to
EpisodeSampler.get_episode().